+++ sancho-src/src/sancho/view/WebBrowserTab$NSBookmark.java
@@ -1,6 +1,6 @@
 package sancho.view;
 
-import sancho.utility.regex.REMatch;
+import java.util.regex.Matcher;
 import org.eclipse.jface.action.Action;
 import sancho.view.utility.SResources;
 
@@ -11,19 +11,23 @@ public class WebBrowserTab$NSBookmark extends Action {
 
    public WebBrowserTab$NSBookmark(WebBrowserTab var1, String var2) {
       this.this$0 = var1;
-      REMatch[] var3 = var1.bookmark_title.getAllMatches(var2);
-      REMatch[] var4 = var1.bookmark_href.getAllMatches(var2);
+      Matcher var3 = var1.bookmark_title.matcher(var2);
+      Matcher var4 = var1.bookmark_href.matcher(var2);
       String var5 = "Unknown";
-      if (var3.length == 1) {
-         int var6 = var3[0].getStartIndex(1);
-         int var7 = var3[0].getEndIndex(1);
-         var5 = var2.substring(var6, var7);
+      if (var3.find()) {
+         int var6 = var3.start(1);
+         int var7 = var3.end(1);
+         if (!var3.find()) {
+            var5 = var2.substring(var6, var7);
+         }
       }
 
-      if (var4.length == 1) {
-         int var8 = var4[0].getStartIndex(1);
-         int var9 = var4[0].getEndIndex(1);
-         this.href = var2.substring(var8, var9);
+      if (var4.find()) {
+         int var8 = var4.start(1);
+         int var9 = var4.end(1);
+         if (!var4.find()) {
+            this.href = var2.substring(var8, var9);
+         }
       }
 
       this.setText(var1.formatTitle(var5));