+++ sancho-src/src/sancho/view/server/ServerViewFrame$HTTPValidator.java
@@ -1,21 +1,21 @@
 package sancho.view.server;
 
-import sancho.utility.regex.RE;
-import sancho.utility.regex.REException;
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
 import org.eclipse.jface.dialogs.IInputValidator;
 import sancho.view.utility.SResources;
 
 class ServerViewFrame$HTTPValidator implements IInputValidator {
-   static RE regex;
+   static Pattern regex;
 
    public String isValid(String var1) {
-      return regex.isMatch(var1) ? null : SResources.getString("l.invalidInput");
+      return regex.matcher(var1).find() ? null : SResources.getString("l.invalidInput");
    }
 
    static {
       try {
-         regex = new RE("http(s)?://\\S*");
-      } catch (REException var1) {
+         regex = Pattern.compile("http(s)?://\\S*");
+      } catch (PatternSyntaxException var1) {
       }
    }
 }